refactor: add core/ and resource_server/ packages (move shared plumbing)
Begin the role-based code namespacing. Introduce oauth2_provider/core/
(shared plumbing) and oauth2_provider/resource_server/ (RS role), and move
the first modules into them with backward-compatible import shims at the old
top-level paths (each emits a DeprecationWarning; removal targeted for 4.0):

- core/: compat, exceptions, http, scopes, signals, utils, checks
- resource_server/: www_authenticate

First-party imports are repointed to the new canonical paths so the library
never triggers its own deprecation warnings. Behavior is unchanged; the full
non-e2e suite passes (996) and app startup emits no self-warnings.

Migration-frozen modules (models, generators, validators) and the central
settings module intentionally stay put.
