+++ oauth2_provider/authorization_server/__init__.py
@@ -1,0 +1,3 @@
+"""Authorization Server role (the OAuth2/OIDC provider side): issuing
+authorization and tokens, client registration, and the OpenID Connect Provider
+identity layer."""
