+++ oauth2_provider/authorization_server/oidc/__init__.py
@@ -1,0 +1,1 @@
+"""OpenID Connect Provider (OP) facet of the Authorization Server."""
