+++ oauth2_provider/compat.py
@@ -1,15 +1,17 @@
+"""Backward-compatible import shim.
+
+``oauth2_provider.compat`` moved to ``oauth2_provider.core.compat`` when the package was reorganized by
+OAuth2 role. Importing from this old path still works but is deprecated and will
+be removed in django-oauth-toolkit 4.0.
 """
-The `compat` module provides support for backwards compatibility with older
-versions of Django and Python.
-"""
 
-try:
-    # Django 5.1 introduced LoginRequiredMiddleware, and login_not_required decorator
-    from django.contrib.auth.decorators import login_not_required
-except ImportError:
+import warnings
 
-    def login_not_required(view_func):
-        return view_func
+from oauth2_provider.core.compat import *  # noqa: F401,F403
 
-
-__all__ = ["login_not_required"]
+warnings.warn(
+    "oauth2_provider.compat has moved to oauth2_provider.core.compat. The old import path is "
+    "deprecated and will be removed in django-oauth-toolkit 4.0.",
+    DeprecationWarning,
+    stacklevel=2,
+)
