+++ oauth2_provider/core/__init__.py
@@ -1,0 +1,2 @@
+"""Core / shared plumbing used across the authorization-server and
+resource-server roles."""
