+++ oauth2_provider/core/signals.py
@@ -1,0 +1,4 @@
+from django.dispatch import Signal
+
+
+app_authorized = Signal()  # providing_args=["request", "token"]
